[idex] continued revision with signature

+++ xchange-idex/src/main/java/org/knowm/xchange/idex/IdexCurrencyMeta.kt
@@ -4,4 +4,4 @@ import org.knowm.xchange.dto.meta.*
 import java.math.*
 
 class IdexCurrencyMeta(scale: Int, withdrawalFee: BigDecimal?, val address: String,
-                       val name: String, val decimals: Int) : CurrencyMetaData(scale, withdrawalFee)
+                       val name: String, val decimals: BigInteger) : CurrencyMetaData(scale, withdrawalFee)
